+++ src/connect/connection-response.cc
@@ -70,6 +70,9 @@ void ConnectionResponse::serialize(std::string& output) const
 #ifdef RESIDUE_SPECIAL_EDITION
         ss << "-SE";
 #endif
+#ifdef RESIDUE_DEBUG
+        ss << "-debug";
+#endif
         serverInfo.insert(std::make_pair("version", ss.str()));
         root["server_info"] = serverInfo;
     }