+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## [Unreleased]
+
 ## [1.2.3] - 31-01-2018
 ### Updates
 - Removed licensing